# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Category:
    properties:
      CategoryName:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Ingredient:
    properties:
      IngredientName:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Recipe:
    properties:
      Categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      CookTime:
        type: integer
      Description:
        type: string
      DifficultyLevel:
        type: integer
      Ingredients:
        items:
          $ref: '#/definitions/models.Ingredient'
        type: array
      RecipeName:
        type: string
      ServingCount:
        type: integer
      Tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      imageName:
        type: string
      updatedAt:
        type: string
    type: object
  models.Tag:
    properties:
      TagName:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      updatedAt:
        type: string
    type: object
info:
  contact:
    name: Ian Hulsbus
    url: https://github.com/ihulsbus/cookbook
  license:
    name: GNU Affero General Public License v3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
paths:
  /ingredient:
    get:
      description: Returns a JSON array of all available ingredients
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Ingredient'
            type: array
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: Any error
          schema:
            type: string
      summary: Get a list of all available ingredients
      tags:
      - ingredients
    post:
      consumes:
      - application/json
      description: Create a new ingredient and return its JSON object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ingredient'
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: Any error
          schema:
            type: string
      summary: Create an ingredient
      tags:
      - ingredients
  /ingredient/{id}:
    get:
      consumes:
      - application/json
      description: Returns a JSON array of all available ingredients
      parameters:
      - description: Ingredient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Ingredient'
            type: array
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: Any error
          schema:
            type: string
      summary: Get a list of all available ingredients
      tags:
      - ingredients
  /recipe:
    get:
      description: Returns a JSON array of all available recipes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Recipe'
            type: array
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: Any error
          schema:
            type: string
      summary: Get a list of all available recipes
      tags:
      - recipes
    post:
      consumes:
      - application/json
      description: Creates a new recipe and returns a JSON object of the created recipe
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: Any error
          schema:
            type: string
      summary: Create a recipe
      tags:
      - recipes
  /recipe/{id}:
    delete:
      description: Deletes a recipe. Returns a simple http status code
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: Any error
          schema:
            type: string
      summary: Delete a recipe
      tags:
      - recipes
    get:
      description: Returns a JSON object of a single recipe
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: Any error
          schema:
            type: string
      summary: Get a single recipes
      tags:
      - recipes
    put:
      consumes:
      - application/json
      description: Updates a single recipe and return the JSON object of the updated
        recipe
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: Any error
          schema:
            type: string
      summary: Update a recipe
      tags:
      - recipes
  /recipe/{id}/cover:
    post:
      consumes:
      - image/jpeg
      description: Upload a recipe image used in the frontend. Returns a simple http
        status code
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: Any error
          schema:
            type: string
      summary: Upload a recipe image
      tags:
      - recipes
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Bearer Token
    type: apiKey
swagger: "2.0"